Creating 2 interceptor services (#16)
* logging


* Adding helm delete functionality to makefile


* Splitting up functionality, logs and more


* checking error fetching services


* Using the proper service name


* enabling RBAC on endpoints so the scaler can access interceptor queue sizes


* logs and TODOs


* using IPs instead of hostnames in external scaler


* backing off polling interval


* logging on external scaler startup


* removing superfluous logging in scaler


* port and updating namespaces


* the right namespace


* adding grpc reflection


* Makefile targets for creating/deleting example HTTPScaledObject


* Removing namespace from example HTTPScaledObject file

it is specified when you add it from the make target


* adding keda creation logic to Makefile (and install doc)


* helm delete keda target


* Passing queue pinger around as a pointer

Otherwise the queue lengths are always reported as 0
