+++ pkg/k8s/deployment.go
@@ -16,7 +16,13 @@ func DeleteDeployment(name string, cl k8sappsv1.DeploymentInterface) error {
 // with the given name and the given image. This does not actually create
 // the deployment in the cluster, it just creates the deployment object
 // in memory
-func NewDeployment(name, image string, ports []int32, env []corev1.EnvVar) *appsv1.Deployment {
+func NewDeployment(
+	name,
+	image string,
+	ports []int32,
+	env []corev1.EnvVar,
+	labels map[string]string,
+) *appsv1.Deployment {
 	containerPorts := make([]corev1.ContainerPort, len(ports))
 	for i, port := range ports {
 		containerPorts[i] = corev1.ContainerPort{
@@ -29,16 +35,16 @@ func NewDeployment(name, image string, ports []int32, env []corev1.EnvVar) *apps
 		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:   name,
-			Labels: labels(name),
+			Labels: labels,
 		},
 		Spec: appsv1.DeploymentSpec{
 			Selector: &metav1.LabelSelector{
-				MatchLabels: labels(name),
+				MatchLabels: labels,
 			},
 			Replicas: int32P(1),
 			Template: corev1.PodTemplateSpec{
 				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels(name),
+					Labels: labels,
 				},
 				Spec: corev1.PodSpec{
 					Containers: []corev1.Container{
