+++ pkg/k8s/scaledobject.go
@@ -36,12 +36,12 @@ func NewScaledObject(name, deploymentName, scalerAddress string) *unstructured.U
 			"kind":       "ScaledObject",
 			"metadata": map[string]interface{}{
 				"name":   name,
-				"labels": labels(name),
+				"labels": Labels(name),
 			},
 			"spec": map[string]interface{}{
 				"minReplicaCount": 0,
 				"maxReplicaCount": 1000,
-				"pollingInterval": 1,
+				"pollingInterval": 250,
 				"scaleTargetRef": map[string]string{
 					"name": deploymentName,
 					// "apiVersion": "apps/v1",
